move read_paper_miriad to own script and delete tests
