+++ simpleDS/tests/test_utils.py
@@ -9,211 +9,12 @@ import copy
 import numpy as np
 import pytest
 from scipy.signal import windows
-from pyuvdata import UVData, utils as uvutils
+from pyuvdata import utils as uvutils
 import pyuvdata.tests as uvtest
 from simpleDS import utils
 from simpleDS.data import DATA_PATH
 from astropy import constants as const
 from astropy import units
-
-
-def test_read_no_file():
-    """Test no file given."""
-    warn_message = [
-        "Antenna positions are not present in the file.",
-        "Antenna positions are not present in the file.",
-        "Ntimes does not match the number of unique " "times in the data",
-        "Xantpos in extra_keywords is a list, array or dict, "
-        "which will raise an error when writing uvfits "
-        "or miriad file types",
-    ]
-
-    pytest.raises(
-        TypeError,
-        uvtest.checkWarnings,
-        utils.read_paper_miriad,
-        func_args=[None, None],
-        func_kwargs={"skip_header": 3, "usecols": [1, 2, 3]},
-        category=UserWarning,
-        nwarnings=len(warn_message),
-        message=warn_message,
-    )
-
-
-def test_load_uv_no_antpos():
-    """Test an Exception is raised when no antpos is provided."""
-    test_file = os.path.join(DATA_PATH, "paper_test_file.uv")
-    warn_message = [
-        "Antenna positions are not present in the file.",
-        "Antenna positions are not present in the file.",
-        "Ntimes does not match the number of unique " "times in the data",
-        "Xantpos in extra_keywords is a list, array or dict, "
-        "which will raise an error when writing uvfits "
-        "or miriad file types",
-    ]
-
-    pytest.raises(
-        ValueError,
-        uvtest.checkWarnings,
-        utils.read_paper_miriad,
-        func_args=[test_file, None],
-        func_kwargs={"skip_header": 3, "usecols": [1, 2, 3]},
-        category=UserWarning,
-        nwarnings=len(warn_message),
-        message=warn_message,
-    )
-
-
-def test_no_antpos_file():
-    """Test an Exception is raised if antpos file does not exist."""
-    test_miriad = os.path.join(DATA_PATH, "paper_test_file.uv")
-    dne_file = os.path.join(DATA_PATH, "not_real_file.txt")
-
-    warn_message = [
-        "Antenna positions are not present in the file.",
-        "Antenna positions are not present in the file.",
-        "Ntimes does not match the number of unique " "times in the data",
-        "Xantpos in extra_keywords is a list, array or dict, "
-        "which will raise an error when writing uvfits "
-        "or miriad file types",
-    ]
-
-    pytest.raises(
-        IOError,
-        uvtest.checkWarnings,
-        utils.read_paper_miriad,
-        func_args=[test_miriad, dne_file],
-        category=UserWarning,
-        nwarnings=len(warn_message),
-        message=warn_message,
-    )
-
-
-def test_antpos_from_file():
-    """Test antpos created from input file."""
-    test_miriad = os.path.join(DATA_PATH, "paper_test_file.uv")
-    test_antpos_file = os.path.join(DATA_PATH, "paper_antpos.txt")
-    warn_message = [
-        "Antenna positions are not present in the file.",
-        "Antenna positions are not present in the file.",
-        "Ntimes does not match the number of unique " "times in the data",
-        "Xantpos in extra_keywords is a list, array or dict, "
-        "which will raise an error when writing uvfits "
-        "or miriad file types",
-    ]
-    pend_dep_message = [
-        "antenna_positions are not defined. "
-        "antenna_positions will be a required parameter in "
-        "version 1.5"
-    ]
-    test_uv = uvtest.checkWarnings(
-        utils.read_paper_miriad,
-        func_args=[test_miriad, test_antpos_file],
-        func_kwargs={"skip_header": 3, "usecols": [1, 2, 3]},
-        category=[UserWarning] * len(warn_message) + [DeprecationWarning],
-        nwarnings=len(warn_message) + 1,
-        message=warn_message + pend_dep_message,
-    )
-
-    read_antpos = np.genfromtxt(test_antpos_file, skip_header=3, usecols=[1, 2, 3])
-
-    test_uvws = np.zeros_like(test_uv.uvw_array)
-    for bl in list(set(test_uv.baseline_array)):
-        baseline_inds = np.where(test_uv.baseline_array == bl)[0]
-        ant_1, ant_2 = test_uv.baseline_to_antnums(bl)
-        uvw = read_antpos[ant_2] - read_antpos[ant_1]
-        test_uvws[baseline_inds, :] = uvw
-
-    assert np.allclose(test_uvws, test_uv.uvw_array)
-
-
-def test_setting_frf_nebw_as_inttime():
-    """Test integration time is set if FRF_NEBW is in extra_keywords."""
-    test_miriad = os.path.join(DATA_PATH, "paper_test_file.uv")
-    test_antpos_file = os.path.join(DATA_PATH, "paper_antpos.txt")
-    warn_message = [
-        "Antenna positions are not present in the file.",
-        "Antenna positions are not present in the file.",
-        "Ntimes does not match the number of unique " "times in the data",
-        "Xantpos in extra_keywords is a list, array or dict, "
-        "which will raise an error when writing uvfits "
-        "or miriad file types",
-    ]
-    pend_dep_message = [
-        "antenna_positions are not defined. "
-        "antenna_positions will be a required parameter in "
-        "version 1.5"
-    ]
-    test_uv = uvtest.checkWarnings(
-        utils.read_paper_miriad,
-        func_args=[test_miriad, test_antpos_file],
-        func_kwargs={"skip_header": 3, "usecols": [1, 2, 3]},
-        category=[UserWarning] * len(warn_message) + [DeprecationWarning],
-        nwarnings=len(warn_message) + 1,
-        message=warn_message + pend_dep_message,
-    )
-    frf_nebw_array = np.ones_like(test_uv.integration_time)
-    frf_nebw_array *= test_uv.extra_keywords["FRF_NEBW"]
-    assert np.allclose(frf_nebw_array, test_uv.integration_time)
-
-
-def test_paper_test_file_requires_bl_conjugation():
-    """Test the input paper test file has baselines which require conjugation."""
-    test_miriad = os.path.join(DATA_PATH, "paper_test_file.uv")
-    warn_message = [
-        "Antenna positions are not present in the file.",
-        "Antenna positions are not present in the file.",
-        "Ntimes does not match the number of unique " "times in the data",
-        "Xantpos in extra_keywords is a list, array or dict, "
-        "which will raise an error when writing uvfits "
-        "or miriad file types",
-    ]
-    pend_dep_message = [
-        "antenna_positions are not defined. "
-        "antenna_positions will be a required parameter in "
-        "version 1.5"
-    ]
-
-    uv1 = UVData()
-    assert (
-        uvtest.checkWarnings(
-            uv1.read_miriad,
-            func_args=[test_miriad],
-            category=[UserWarning] * len(warn_message) + [DeprecationWarning],
-            nwarnings=len(warn_message) + 1,
-            message=warn_message + pend_dep_message,
-        )
-        is None
-    )
-    assert np.logical_not(np.all(uv1.uvw_array[:, 0] > 0))
-
-
-def test_conjugating_baselines_in_paper_read():
-    """Test that baslines are properly cojugated when read with utils."""
-    test_miriad = os.path.join(DATA_PATH, "paper_test_file.uv")
-    test_antpos_file = os.path.join(DATA_PATH, "paper_antpos.txt")
-    warn_message = [
-        "Antenna positions are not present in the file.",
-        "Antenna positions are not present in the file.",
-        "Ntimes does not match the number of unique " "times in the data",
-        "Xantpos in extra_keywords is a list, array or dict, "
-        "which will raise an error when writing uvfits "
-        "or miriad file types",
-    ]
-    pend_dep_message = [
-        "antenna_positions are not defined. "
-        "antenna_positions will be a required parameter in "
-        "version 1.5"
-    ]
-    test_uv = uvtest.checkWarnings(
-        utils.read_paper_miriad,
-        func_args=[test_miriad, test_antpos_file],
-        func_kwargs={"skip_header": 3, "usecols": [1, 2, 3]},
-        category=[UserWarning] * len(warn_message) + [DeprecationWarning],
-        nwarnings=len(warn_message) + 1,
-        message=warn_message + pend_dep_message,
-    )
-    assert np.all(test_uv.uvw_array[:, 0] > 0)
 
 
 def test_get_data_array():
